feat: implement validation in sensors endpoints and add sensor-management feature with template using Dustjs and vanilla JS

+++ services/api/src/application/use-cases/sensors/delete-sensor.use-case.ts
@@ -1,4 +1,5 @@
 import { SensorRepositoryInstance, SensorRepositoryType } from "../../../infrastructure/database/postgres/repositories/sensor.repository";
+import { NotFoundException } from "../../../infrastructure/http/exceptions/NotFoundException";
 
 export class DeleteSensorUseCase {
 
@@ -9,6 +10,10 @@ export class DeleteSensorUseCase {
     }
 
     async execute(id: string) {
-        return await this._sensorRepository.delete(id);
+        const sensor = await this._sensorRepository.getById(id);
+        if (!sensor) {
+            throw new NotFoundException(`Sensor not found`);
+        }
+        await this._sensorRepository.delete(id);
     }
 }
